+++ gulpfile.js
@@ -33,25 +33,14 @@ gulp.task('clean', function() {
 
 // uglify task
 gulp.task('js', function() {
-    // main app js file
-    var datatables_src = [
+    
+    var combined_src = [
+        './src/app.js',
         './src/_vendor/ui.bootstrap/src/pagination/pagination.js',
         './build/partials/datatables/*.js',
         './build/_vendor/**/*.js',
-        './src/datatables/app.js',
         './src/datatables/**/*.js',
-        '!./src/**/*.spec.js'
-    ];
-
-    var validation_src = [
-        './build/partials/validation/*.js',
-        './src/validation/app.js',
         './src/validation/**/*.js',
-        '!./src/**/*.spec.js'
-    ];
-
-    var combined_src = [
-        './src/app.js',
         '!./src/**/*.spec.js'
     ]
 
@@ -59,16 +48,12 @@ gulp.task('js', function() {
         .pipe(replace(/\{version\}/g, pkg.version))
         .pipe(gulp.dest('./build/'));
 
-    build('oDirectives', [combined_src, datatables_src, validation_src]);
-    build('oDirectives.datatables', [datatables_src]);
-    build('oDirectives.validation', [validation_src]);
-
+    build('oDirectives', [combined_src]);
 
     function build(outputFileName, sources) {
         var combined_build = Array.prototype.concat.apply([], sources)
         combined_build.unshift('./build/build_head.js');
         
-
         //console.log(combined_build);
         //
         var the_source = gulp.src(combined_build, base);
@@ -91,7 +76,7 @@ gulp.task('templatify', function() {
     gulp
         .src("./src/_vendor/ui.bootstrap/template/pagination/*.html")
         .pipe(ngHtml2Js({
-            moduleName: "envoc.directives.partials",
+            moduleName: "envoc",
             prefix: "template/pagination/"
         }))
         .pipe(gulp.dest("./build/_vendor/templates"));
@@ -99,7 +84,7 @@ gulp.task('templatify', function() {
     return gulp
         .src("./src/**/*.tmpl.html")
         .pipe(ngHtml2Js({
-            moduleName: "envoc.directives.partials",
+            moduleName: "envoc",
             prefix: "/oTemplates/"
         }))
         .pipe(gulp.dest("./build/partials"));