+++ src/datatables/app.js
@@ -1,4 +1,0 @@
-angular.module('envoc.directives.datatables', [
-    'envoc.directives.partials',
-    'ui.bootstrap.pagination'
-]);